% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Adep}
\alias{Adep}
\title{Example acoustic deployment field-notebook}
\format{
A data frame with 2 rows and 14 variables:
\describe{
\item{deploymentID}{Unique identifier of the deployment.}
\item{longitude}{Longitude in decimal degrees (WGS84).}
\item{latitude}{Latitude in decimal degrees (WGS84).}
\item{locationID}{Identifier of the deployment location.}
\item{startDate}{Deployment start date.}
\item{startTime}{Deployment start time.}
\item{endDate}{Deployment end date.}
\item{endTime}{Deployment end time.}
\item{deviceID}{Identifier of the recording device.}
\item{deviceModel}{Manufacturer and model of the recording device.}
\item{samplingFrequency}{Sampling frequency of the recordings (Hz).}
\item{bitDepth}{Bit depth of the recordings.}
\item{channels}{Number of audio channels.}
\item{setupBy}{Name or identifier of the person/organization that deployed
the device.}
}
}
\usage{
Adep
}
\description{
Example deployment notebook for an acoustic (audio) survey, used in the
acoustic vignette. One row per device deployment. Coordinates are random
points within the inland Izu Peninsula (Shizuoka, Japan).
}
\seealso{
\link{Aobs}
}
\keyword{datasets}
