% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDP_functions.r
\name{create_deployments}
\alias{create_deployments}
\title{Create deployments}
\usage{
create_deployments(
  deploymentID,
  latitude,
  longitude,
  deploymentStart = NULL,
  deploymentStart_date = NULL,
  deploymentStart_time = NULL,
  deploymentEnd = NULL,
  deploymentEnd_date = NULL,
  deploymentEnd_time = NULL,
  locationID = NULL,
  locationName = NULL,
  coordinateUncertainty = NULL,
  setupBy = NULL,
  cameraID = NULL,
  cameraModel = NULL,
  cameraDelay = NULL,
  cameraHeight = NULL,
  cameraDepth = NULL,
  cameraTilt = NULL,
  cameraHeading = NULL,
  detectionDistance = NULL,
  timestampIssues = NULL,
  baitUse = NULL,
  featureType = NULL,
  habitat = NULL,
  deploymentGroups = NULL,
  deploymentTags = NULL,
  deploymentComments = NULL,
  tz = "Japan"
)
}
\value{
A tibble of deployments.
}
\description{
\code{create_deployments} creates a table of deployments.
}
