% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_report.R
\name{ctdp_is_valid}
\alias{ctdp_is_valid}
\title{Did a validation pass (no errors)?}
\usage{
ctdp_is_valid(issues)
}
\arguments{
\item{issues}{An issue table.}
}
\value{
\code{TRUE} if there are no \code{"error"} severity rows.
}
\description{
Did a validation pass (no errors)?
}
