% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datapackageIdata}
\alias{datapackageIdata}
\title{Camtrap DP objects have been created from Idep and Iobs}
\description{
Camtrap DP objects have been created from Idep and Iobs
}
\keyword{camtrapDP}
\keyword{data}
