% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datapackageVdata}
\alias{datapackageVdata}
\title{Camtrap DP objects have been created from Vdep and Vobs}
\description{
Camtrap DP objects have been created from Vdep and Vobs
}
\references{
Originally data can be used from https://doi.org/10.34462/0002000233
}
\keyword{camtrapDP}
\keyword{data}
