% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformance.R
\name{.ctdp_allowed_constraints}
\alias{.ctdp_allowed_constraints}
\title{Constraint keys that are valid for a given field type}
\usage{
.ctdp_allowed_constraints(type)
}
\arguments{
\item{type}{A Frictionless field type.}
}
\description{
Constraint keys that are valid for a given field type
}
\keyword{internal}
