% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ctdp_date_format}
\alias{.ctdp_date_format}
\alias{.ctdp_time_format}
\title{Default Camtrap DP date / time formats}
\usage{
.ctdp_date_format()

.ctdp_time_format()
}
\description{
Default Camtrap DP date / time formats
}
\keyword{internal}
