% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ctdp_datetime_format}
\alias{.ctdp_datetime_format}
\title{Default Camtrap DP datetime format}
\usage{
.ctdp_datetime_format()
}
\description{
The Frictionless \code{datetime} format used by the Camtrap DP table schemas.
}
\keyword{internal}
