% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformance.R
\name{.ctdp_frictionless_types}
\alias{.ctdp_frictionless_types}
\title{Field types supported by the Frictionless Table Schema specification}
\usage{
.ctdp_frictionless_types()
}
\description{
Field types supported by the Frictionless Table Schema specification
}
\keyword{internal}
