% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ctdp_infer_field}
\alias{.ctdp_infer_field}
\title{Build a Frictionless field definition for a custom column}
\usage{
.ctdp_infer_field(name, x)
}
\arguments{
\item{name}{Column name.}

\item{x}{The column vector (used to infer the type).}
}
\description{
Build a Frictionless field definition for a custom column
}
\keyword{internal}
