% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ctdp_infer_type}
\alias{.ctdp_infer_type}
\title{Infer a Frictionless field type from an R vector}
\usage{
.ctdp_infer_type(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A Frictionless \code{type} string.
}
\description{
Infer a Frictionless field type from an R vector
}
\keyword{internal}
