% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ctdp_is_missing}
\alias{.ctdp_is_missing}
\title{Test whether a character value represents a Frictionless "missing value"}
\usage{
.ctdp_is_missing(x, missing_values = "")
}
\arguments{
\item{x}{A character vector.}

\item{missing_values}{Character vector of tokens treated as missing.}
}
\description{
Test whether a character value represents a Frictionless "missing value"
}
\keyword{internal}
