% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformance.R
\name{.ctdp_known_constraints}
\alias{.ctdp_known_constraints}
\title{Constraint keys defined by the Frictionless Table Schema specification}
\usage{
.ctdp_known_constraints()
}
\description{
Constraint keys defined by the Frictionless Table Schema specification
}
\keyword{internal}
