% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{.ctdp_normalize_mapping}
\alias{.ctdp_normalize_mapping}
\title{Normalise a mapping into a tibble with \code{source} and \code{target} columns}
\usage{
.ctdp_normalize_mapping(mapping)
}
\description{
Normalise a mapping into a tibble with \code{source} and \code{target} columns
}
\keyword{internal}
