% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frictionless.R
\name{.ctdp_resolve_frictionless_script}
\alias{.ctdp_resolve_frictionless_script}
\title{Resolve the path to the Python helper script}
\usage{
.ctdp_resolve_frictionless_script(script = NULL)
}
\description{
Resolve the path to the Python helper script
}
\keyword{internal}
