% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ctdp_sub_version}
\alias{.ctdp_sub_version}
\title{Substitute the \verb{<version>} placeholder in a URL template}
\usage{
.ctdp_sub_version(template, version)
}
\arguments{
\item{template}{A URL containing \verb{<version>}.}

\item{version}{The Camtrap DP version string.}
}
\description{
Substitute the \verb{<version>} placeholder in a URL template
}
\keyword{internal}
