% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a, b}{Values.}
}
\description{
Returns \code{a} unless it is \code{NULL} (or zero-length), otherwise \code{b}.
}
\keyword{internal}
