% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Idep}
\alias{Idep}
\title{Deployment data of images which collected by 10 camera-trap (dummy data)}
\description{
Deployment data of images which collected by 10 camera-trap (dummy data)
}
\keyword{data}
