% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Iobs}
\alias{Iobs}
\title{Observation data of image and media-based observation which collected by 10 camera-trap (dummy data)}
\description{
Observation data of image and media-based observation which collected by 10 camera-trap (dummy data)
}
\keyword{data}
