% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Vobs}
\alias{Vobs}
\title{Observation data of videos which collected by a camera-trap}
\description{
Observation data of videos which collected by a camera-trap
}
\references{
Originally data can be used from https://doi.org/10.34462/0002000233
}
\keyword{data}
