% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDP_functions.r
\name{create_media}
\alias{create_media}
\title{Create media}
\usage{
create_media(
  mediaID,
  deploymentID,
  timestamp = NULL,
  timestamp_date = NULL,
  timestamp_time = NULL,
  filePath,
  filePublic,
  fileMediatype,
  fileName = NULL,
  captureMethod = NULL,
  exifData = NULL,
  favorite = NULL,
  mediaComments = NULL,
  tz = "Japan",
  omitduplicate = TRUE
)
}
\value{
A tibble of media.
}
\description{
\code{create_media} creates a table of media.
}
