% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDP_functions.r
\name{R6_CamtrapDP}
\alias{R6_CamtrapDP}
\title{R6 class representing Camtrap DP}
\description{
R6 class including metadata, deployments, media and observations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{resources}}{is the package data resources}

\item{\code{profile}}{of the resource}

\item{\code{name}}{Identifier of the resource}

\item{\code{id}}{A property reserved for globally unique identifiers e.g., UUID and DOI}

\item{\code{created}}{The datetime on which this Data Package was created}

\item{\code{title}}{Title of this Data Package}

\item{\code{contributors}}{The people or organizations who contributed to this Data Package}

\item{\code{description}}{Description of this Data Package}

\item{\code{version}}{The version of this Data Package}

\item{\code{keywords}}{Keywords of this Data Package}

\item{\code{image}}{A URL or Path of an image for this Data Package}

\item{\code{homepage}}{A URL for the home on the web that is related to this Data Package}

\item{\code{sources}}{A row sources for this Data Package}

\item{\code{licenses}}{The licenses under which the Data Package is provided}

\item{\code{bibliographicCitation}}{A bibliographical reference for the resource}

\item{\code{project}}{Camera trap project or study that originated this Data package}

\item{\code{coordinatePrecision}}{Least precise coordinate precision of the deployments.latitude and deployments.longitude}

\item{\code{spatial}}{Spatial coverage of this Data Package, expressed as GeoJSON}

\item{\code{temporal}}{Temporal coverage of this Data Package}

\item{\code{taxonomic}}{Taxonomic coverage of this Data Package, based on the unique observations.scientificName}

\item{\code{relatedIdentifiers}}{Identifiers of resources related to this Data Package}

\item{\code{references}}{List of references related to this Data Package}

\item{\code{directory}}{Directory of this Data Package}

\item{\code{data}}{Observation, Media and Deployments which consist this Data Package}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CamtrapDP-new}{\code{R6_CamtrapDP$new()}}
\item \href{#method-CamtrapDP-update_created}{\code{R6_CamtrapDP$update_created()}}
\item \href{#method-CamtrapDP-set_properties}{\code{R6_CamtrapDP$set_properties()}}
\item \href{#method-CamtrapDP-set_deployments}{\code{R6_CamtrapDP$set_deployments()}}
\item \href{#method-CamtrapDP-set_media}{\code{R6_CamtrapDP$set_media()}}
\item \href{#method-CamtrapDP-set_observations}{\code{R6_CamtrapDP$set_observations()}}
\item \href{#method-CamtrapDP-set_custom}{\code{R6_CamtrapDP$set_custom()}}
\item \href{#method-CamtrapDP-add_contributors}{\code{R6_CamtrapDP$add_contributors()}}
\item \href{#method-CamtrapDP-add_sources}{\code{R6_CamtrapDP$add_sources()}}
\item \href{#method-CamtrapDP-add_license}{\code{R6_CamtrapDP$add_license()}}
\item \href{#method-CamtrapDP-set_project}{\code{R6_CamtrapDP$set_project()}}
\item \href{#method-CamtrapDP-set_st}{\code{R6_CamtrapDP$set_st()}}
\item \href{#method-CamtrapDP-set_taxon}{\code{R6_CamtrapDP$set_taxon()}}
\item \href{#method-CamtrapDP-add_relatedIdentifiers}{\code{R6_CamtrapDP$add_relatedIdentifiers()}}
\item \href{#method-CamtrapDP-add_references}{\code{R6_CamtrapDP$add_references()}}
\item \href{#method-CamtrapDP-out_camtrapdp}{\code{R6_CamtrapDP$out_camtrapdp()}}
\item \href{#method-CamtrapDP-import_metadata}{\code{R6_CamtrapDP$import_metadata()}}
\item \href{#method-CamtrapDP-clone}{\code{R6_CamtrapDP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-new"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-new}{}}}
\subsection{Method \code{new()}}{
Creates new instance of R6_CamtrapDP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$new(tz = "Japan", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tz}}{Time zone.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-update_created"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-update_created}{}}}
\subsection{Method \code{update_created()}}{
Updates timestamp of R6_CamtrapDP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$update_created(tz = "Japan")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tz}}{Time zone.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-set_properties"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-set_properties}{}}}
\subsection{Method \code{set_properties()}}{
Sets properties of R6_CamtrapDP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$set_properties(
  directory = getwd(),
  name = NULL,
  id = NULL,
  description = NULL,
  profile =
    "https://raw.githubusercontent.com/tdwg/camtrap-dp/<version>/camtrap-dp-profile.json",
  version = "1.0.1",
  keywords = NULL,
  image = NULL,
  homepage = NULL,
  bibliographicCitation = NULL,
  coordinatePrecision = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{directory}}{Directory of datapackage.}

\item{\code{name}}{Identifier of the resource}

\item{\code{id}}{A property reserved for globally unique identifiers e.g., UUID and DOI}

\item{\code{description}}{Description of this Data Package}

\item{\code{profile}}{Profile of the resource}

\item{\code{version}}{The version of this Data Package}

\item{\code{keywords}}{Keywords of this Data Package}

\item{\code{image}}{A URL or Path of an image for this Data Package}

\item{\code{homepage}}{A URL for the home on the web that is related to this Data Package}

\item{\code{bibliographicCitation}}{A bibliographical reference for the resource}

\item{\code{coordinatePrecision}}{Least precise coordinate precision of the deployments.latitude and deployments.longitude}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-set_deployments"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-set_deployments}{}}}
\subsection{Method \code{set_deployments()}}{
Sets deployments of R6_CamtrapDP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$set_deployments(
  data,
  path = "deployments.csv",
  profile = "tabular-data-resource",
  format = "csv",
  mediatype = "text/csv",
  encoding = "utf-8",
  schema =
    "https://raw.githubusercontent.com/tdwg/camtrap-dp/<version>/deployments-table-schema.json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Deployments dataset.}

\item{\code{path}}{Path or URL to the data file}

\item{\code{profile}}{Profile of deployments}

\item{\code{format}}{Format of deployments data}

\item{\code{mediatype}}{Media type of deployments data}

\item{\code{encoding}}{Encpding of deployments data}

\item{\code{schema}}{URL of the used Camtrap DP Table Schema version}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-set_media"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-set_media}{}}}
\subsection{Method \code{set_media()}}{
Sets media of R6_CamtrapDP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$set_media(
  data,
  path = "media.csv",
  profile = "tabular-data-resource",
  format = "csv",
  mediatype = "text/csv",
  encoding = "utf-8",
  schema =
    "https://raw.githubusercontent.com/tdwg/camtrap-dp/<version>/media-table-schema.json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Media dataset.}

\item{\code{path}}{Path or URL to the data file}

\item{\code{profile}}{Profile of media data}

\item{\code{format}}{Format of media data}

\item{\code{mediatype}}{Mediatype of media data}

\item{\code{encoding}}{Encoding of media data}

\item{\code{schema}}{URL of the used Camtrap DP Table Schema version}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-set_observations"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-set_observations}{}}}
\subsection{Method \code{set_observations()}}{
Sets observations of R6_CamtrapDP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$set_observations(
  data,
  path = "observations.csv",
  profile = "tabular-data-resource",
  format = "csv",
  mediatype = "text/csv",
  encoding = "utf-8",
  schema =
    "https://raw.githubusercontent.com/tdwg/camtrap-dp/<version>/observations-table-schema.json"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Observations dataset.}

\item{\code{path}}{Path of URL to the data file}

\item{\code{profile}}{Profile of observations}

\item{\code{format}}{Format of observation data}

\item{\code{mediatype}}{Mediatype of observation data}

\item{\code{encoding}}{Encoding of observation data}

\item{\code{schema}}{URL of the used Camtrap DP Table Schema version}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-set_custom"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-set_custom}{}}}
\subsection{Method \code{set_custom()}}{
Sets custom data of R6_CamtrapDP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$set_custom(name, description, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of dataset.}

\item{\code{description}}{Description of dataset.}

\item{\code{data}}{Custom dataset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-add_contributors"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-add_contributors}{}}}
\subsection{Method \code{add_contributors()}}{
Adds contributors of R6_CamtrapDP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$add_contributors(contrib_table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{contrib_table}}{data frame or tibble of contributors including title, email, path, role and organizaiton.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-add_sources"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-add_sources}{}}}
\subsection{Method \code{add_sources()}}{
Add sources of R6_CamtrapDP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$add_sources(title, path = NULL, email = NULL, version = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{Title of sources.}

\item{\code{path}}{Path or URL to the source}

\item{\code{email}}{An email address}

\item{\code{version}}{The version of this Data Package}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-add_license"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-add_license}{}}}
\subsection{Method \code{add_license()}}{
Add license of R6_CamtrapDP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$add_license(name, scope, path = NULL, title = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of license.}

\item{\code{scope}}{scope of license ("data" or "media").}

\item{\code{path}}{A URL or path to the details of license}

\item{\code{title}}{A title of license}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-set_project"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-set_project}{}}}
\subsection{Method \code{set_project()}}{
Sets project of R6_CamtrapDP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$set_project(
  title,
  samplingDesign,
  captureMethod,
  individualAnimals,
  observationLevel,
  id = NULL,
  acronym = NULL,
  description = NULL,
  path = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{Title of project.}

\item{\code{samplingDesign}}{Sampling design. Either "(simpleRandom|systematicRandom|clusteredRandom|experimental|targeted|opportunistic)".}

\item{\code{captureMethod}}{Capture method. Either "(activityDetection|timeLapse)".}

\item{\code{individualAnimals}}{Logical indicating whether the individuals are recognized.}

\item{\code{observationLevel}}{Observation level. Either "(media|event)".}

\item{\code{id}}{Unique identifier of the project}

\item{\code{acronym}}{Project acronym}

\item{\code{description}}{Description of the project}

\item{\code{path}}{Project website}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-set_st"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-set_st}{}}}
\subsection{Method \code{set_st()}}{
Sets spatial and temporal of R6_CamtrapDP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$set_st()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-set_taxon"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-set_taxon}{}}}
\subsection{Method \code{set_taxon()}}{
Sets taxonomic of R6_CamtrapDP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$set_taxon(taxonDB = "itis", taxonDBurl = "https://www.itis.gov/")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{taxonDB}}{Name of taxon data base passed to \code{taxadb::get_ids()}.}

\item{\code{taxonDBurl}}{An URL to taxon data pase}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-add_relatedIdentifiers"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-add_relatedIdentifiers}{}}}
\subsection{Method \code{add_relatedIdentifiers()}}{
Add relatedIdentifiers of R6_CamtrapDP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$add_relatedIdentifiers(
  relationType,
  relatedIdentifier,
  relatedIdentifierType,
  resourceTypeGeneral = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{relationType}}{Type of relation. Either "^(IsCitedBy|Cites|IsSupplementTo|IsSupplementedBy|IsContinuedBy|Continues|IsNewVersionOf|IsPreviousVersionOf|IsPartOf|HasPart|IsPublishedIn|IsReferencedBy|References|IsDocumentedBy|Documents|IsCompiledBy|Compiles|IsVariantFormOf|IsOriginalFormOf|IsIdenticalTo|HasMetadata|IsMetadataFor|Reviews|IsReviewedBy|IsDerivedFrom|IsSourceOf|Describes|IsDescribedBy|HasVersion|IsVersionOf|Requires|IsRequiredBy|Obsoletes|IsObsoletedBy)$".}

\item{\code{relatedIdentifier}}{Related identifier.}

\item{\code{relatedIdentifierType}}{Type of related identifier. Either "^(ARK|arXiv|bibcode|DOI|EAN13|EISSN|Handle|IGSN|ISBN|ISSN|ISTC|LISSN|LSID|PMID|PURL|UPC|URL|URN|w3id)$".}

\item{\code{resourceTypeGeneral}}{General type of the related resource}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-add_references"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-add_references}{}}}
\subsection{Method \code{add_references()}}{
Add references of R6_CamtrapDP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$add_references(reference)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reference}}{Reference of data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-out_camtrapdp"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-out_camtrapdp}{}}}
\subsection{Method \code{out_camtrapdp()}}{
Exports \code{camtrapdp} object and datapackage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$out_camtrapdp(write = FALSE, directory = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{write}}{If TRUE, datapackage is written to \code{directory}.}

\item{\code{directory}}{Path to the directory where new data packages will save}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{camtrapdp} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-import_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-import_metadata}{}}}
\subsection{Method \code{import_metadata()}}{
Import metadata from a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$import_metadata(metadata0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata0}}{List of metadata.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CamtrapDP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CamtrapDP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_CamtrapDP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
