% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Vdep}
\alias{Vdep}
\title{Deployment data of videos which collected by a camera-trap}
\description{
Deployment data of videos which collected by a camera-trap
}
\references{
Originally data can be used doi.xxx.xxx
}
\keyword{data}
